feat: changed some styles
Now the message align sets base on what language you're typing with.
Also the background has changed. And border radiuses too.
And now the app gives your message its own specific font base on the language you typed your message with.
If your message is typed with a right-to-left language it'll have an Arabic font
and if it's typed with a left-to-right language it'll have an English font.
Also the messages look way better now with glassmorphism effect on them.
Also background on phone is different from the one on desktop.
The desktop one is horizontal and the phone one is vertical.